% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/QClasses.R
\docType{methods}
\name{AddId,QTable,character-method}
\alias{AddId,QTable,character-method}
\title{Method definition for AddId from QTable}
\usage{
\S4method{AddId}{QTable,character}(object, idName)
}
\arguments{
\item{object:}{QTable to which id attribute should be added}

\item{idName:}{column name from QTable}
}
\value{
QTable: updated QTable
}
\description{
Method definition for AddId from QTable
}

