% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/apply-block.R
\name{apply_block}
\alias{apply_block}
\title{Apply blocking}
\usage{
apply_block(table_a, table_b, func_obj, ..., col_names_a = NULL,
  col_names_b = NULL)
}
\arguments{
\item{table_a,table_b}{Input qtables, table A and table B.}

\item{func_obj}{Blocking function to be applied. It is a R function which is
expected to take two tuples (and other parameters) and return 1 if they
match and 0 if they dont. Note: Currently if the function object's name is
\code{attr_equiv_block} (see \code{\link{show_blockers}}, the parameters
are attribute name from table A to be matched with attribute name from
table B. The blocking function is not invoked for every tuple pair, instead
a join based on attributes is done.}

\item{...}{Parameters relevant to blocking function. For
\code{attr_equiv_block} it is the attribute name from table A and table B.
Currently blocking based on single attribute is supported.}

\item{col_names_a}{(optional) List of column names from table A to be
included in the candidate set.}

\item{col_names_b}{(optional) List of column names from table B to be
incldued in the candidate set.}
}
\value{
A candidate set represented as qtable object. The candidate set
  contains the following: \code{id} attributes from table A and table B,
  optional columns as mentioned in col_names_a, optional columns as mentioned
  in col_names_b.
}
\description{
It takes in two qtables and a blocking function. It applies the blocking
function to tuple pairs from input table and return a candidate set with
potentially matching tuple pairs.
}
\examples{
\dontrun{
  attr1 < "isbn"
  attr2 <- "isbn"
  cand_set <- apply_block(walmart, bowker,
               attr_equiv_block, attr1, attr2, coln
             )
  }
}

