% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qtable.R
\name{check_id}
\alias{check_id}
\title{Check Id attribute}
\usage{
check_id(object, key)
}
\arguments{
\item{object}{qtable object.}

\item{key}{List of column (attribute) names from qtable object}
}
\value{
A numeric value (0/1) indicating whether the list of attribute names can be
  set as key.
}
\description{
Checks whether a list of attributes contain unique values and does not have
any missing values.
}
\examples{
\dontrun{
 status <- check_id(qtable_obj, list("id"))
 }
}

