% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/apply-feature-fns.R
\name{convert_to_feature_vecs}
\alias{convert_to_feature_vecs}
\title{Convert to feature vectors}
\usage{
convert_to_feature_vecs(table_a, table_b, cand_set, fn_list)
}
\arguments{
\item{table_a,table_b}{Base tables represented as qtable objects.}

\item{cand_set}{Labeled data (or candidate set) represented as qtable.}

\item{fn_list}{List of features represented as R functions.}
}
\value{
Returns a qtable with feature values.
}
\description{
Given base tables, labeled data (or candidate set) and feature list (R
function). \code{convert_to_feature_vec} computes the features by joining
candidate set to base tables and applying features represented as R
functions.
}
\examples{
\dontrun{
 labeled_feat_vec <- convert_to_feature_vecs(walmart, bowker,
                       labeled_data,
                       feat_list
                     )
}
}

