% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create-feature-fns.R
\name{create_features}
\alias{create_features}
\title{Create features}
\usage{
create_features(table_a, table_b, attr_corres = NULL,
  allowed_simlist = NULL, allowed_toklist = NULL)
}
\arguments{
\item{table_a,table_b}{Input qtable objects.}

\item{attr_corres}{List of character vectors capturing the attribute
correspondence between table_a and table_b. If set to NULL, attribute
correspondence is done based on attribute names.}

\item{allowed_simlist}{List of strings containing the similarity function
names (a subset of global similarity function list) to be considered for
automatic feature creation. If set to NULL, global similarity function list
(see \code{\link{show_simfuns}}) is considered.}

\item{allowed_toklist}{List of strings containing the tokenizer names (a
  subset of global tokenizer list) to be considered for automatic feature
  creation. If set to NULL, global tokenizer list (see
  \code{\link{show_tokenizers}}) is considered.}
}
\value{
Returns a list of features represented as R functions that takes in
  two tuples and return a numeric value.
}
\description{
Automatically create features based on input qtable's attribute type and its
statistics.
}
\examples{
\dontrun{
feature_list <- create_features(walmart, bowker,
                 attr_corres = list(c("title", "title),c("binding", "binding")),
                 allowed_sim_list = list("jaccard", "lev"),
                 allowed_tok_list = list("tok_qgram")
                 )
}
}

