% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean-data.R
\name{fix}
\alias{fix}
\title{Fix a qtable}
\usage{
fix(x, ...)
}
\arguments{
\item{x}{qtable object}

\item{...}{Relevant paramaters to \code{\link{edit}} function.}
}
\description{
\code{fix} internal invokes \code{\link{edit}} on qtable object and
stores back the updated version of qtable object.
}
\examples{
\\donotrun{
fix(qtable_obj)
}
}

