% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/label-data.R
\name{label_data}
\alias{label_data}
\title{Label data}
\usage{
label_data(sampled_data)
}
\arguments{
\item{sampled_data}{Sampled candidate set represented as qtable.}
}
\value{
qtable object with columns from input qtable along with "label"
  column updated by the user.
}
\description{
Lets the user provide labeled data that can be used for model selection,
training and prediction. \code{label_data} opens a GUI with data from
candidate set and an extra column named "label". The user can manually update
the labels with values 0 to indicate that the tuple pairs donot match and 1
to indicate that the tuple pairs match.
}
\examples{
\dontrun{
labeled_Data <- label_data(sampled_data)
}
}

