% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict-label.R
\name{predict_label}
\alias{predict_label}
\title{Predict labels}
\usage{
predict_label(feat_table, model)
}
\arguments{
\item{feat_table}{Feature vectors represented as qtable,  computed from
labeled data.}

\item{model}{ML model learned from labeled data.}
}
\value{
Returns a qtable, the feature vector from input table is appended
  with "predicted_labels" column that contains the predictions from learned
  model. The "predicted_labels" contains either 0 to indicate non-match or 1
  to indicate match.
}
\description{
Predict labels typically for tuple pairs in candidate set using model learned
from labeled data.
}
\examples{
\dontrun{
 dt_model <- train_model(labeled_feat_vec, method = "rpart")
 labeled_candset <- predict_label(labeled_feat_vec, dt_model)
}
}

