\name{QPackage-package}
\alias{QPackage-package}
\alias{QPackage}
\docType{package}
\title{
Entity Matching package
}
\description{
End-to-end entity matching management system
}
\details{
\tabular{ll}{
Package: \tab QPackage\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-03-22\cr
License: \tab MIT\cr
}
Supports end-to-end workflow for entity matching management system
}
\author{
Pradap KV

Maintainer: Pradap KV <pradap@cs.wisc.edu>
}
\references{

}
\keyword{ package }
\seealso{


}
\examples{

}
