% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/QClasses.R
\docType{class}
\name{QTable-class}
\alias{QTable-class}
\title{Class definition for QTable
Has two attributes: idField and genId
idField : column id name in the table
genId : flag to denote whether Id attribute must be generated.}
\description{
Class definition for QTable
Has two attributes: idField and genId
idField : column id name in the table
genId : flag to denote whether Id attribute must be generated.
}

