% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qtable.R
\name{qtable}
\alias{qtable}
\title{An S4 class to represent qtable}
\usage{
qtable(...)
}
\description{
An S4 class to represent qtable
}
\section{Slots}{

\describe{
\item{\code{key}}{A list of attributes in qtable that uniquely identifies a row.
The key values are unique and do not contain missing values.}
}}

