% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/import-data.R
\name{read_csv}
\alias{read_csv}
\title{Import csv file into qtable}
\usage{
read_csv(..., key = NULL, status = NULL)
}
\arguments{
\item{...}{Relavant parameters to \code{\link{read.csv}}.}

\item{key}{(optional) List of attribute names that can be set as key.}

\item{status}{(optional) Symbol to store return status.}
}
\description{
Reads a file in table format and creates a qtable from it. It uses \code{\link{read.csv}}
to read the contents of csv file.
}
\examples{
\dontrun{
bowker_table <- read_csv("bowker.csv",
                 key = list("id"),
                 status = read_status)
}
}

