% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qtable.R
\name{set_id}
\alias{set_id}
\title{Set Id attribute}
\usage{
set_id(object, key)
}
\arguments{
\item{object}{qtable object.}

\item{key}{List of column (attribute) names from qtable object}
}
\value{
Returns a numeric status (0/1) indicating whether the function
  executed successfully.
}
\description{
Sets a list of attributes as key to a qtable object.
}
\examples{
\dontrun{
status <- set_id(qtable_obj, list("id"))
}
}

