% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/QClasses.R
\docType{methods}
\name{SetId,QTable,character-method}
\alias{SetId,QTable,character-method}
\title{Method definition for CheckId from QTable}
\usage{
\S4method{SetId}{QTable,character}(object, idCol)
}
\arguments{
\item{object:}{QTable to which id attribute should be added}

\item{idCol:}{column name from QTable}
}
\value{
QTable: updated QTable
}
\description{
Method definition for CheckId from QTable
}

