% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{show_simfuns}
\alias{show_simfuns}
\title{Show similarity functions}
\usage{
show_simfuns()
}
\description{
Display global list of similarity functions.
}
\examples{
\dontrun{
 show_simfuns()
}
}

