% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/train-model.R
\name{train_model}
\alias{train_model}
\title{Train model}
\usage{
train_model(feat_table, method = NULL, ...)
}
\arguments{
\item{feat_table}{Feature vectors represented as qtable,  computed from labeled data.}

\item{method}{Learning model to be applied represented as string. Currently
  seven learning models are supported  (see \code{\link{show_learners}}). The
  method should be one of the following: "svm", "rpart", "randomforest",
  "nnet", "bagging", "boosting", "naivebayes".}

\item{...}{Optional parameters relevant to learning model.}
}
\value{
Returns trained model.
}
\description{
Train supervised learning model using features from labeled data.
}
\examples{
\dontrun{
 dt_model <- train_model(labeled_feat_vec, method = "rpart")
}
}

