% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tuple.R
\name{tuple}
\alias{tuple}
\title{Tuple constructor
Tuple contructor that can take in a R named list and return Tuple object}
\usage{
tuple(...)
}
\description{
Tuple constructor
Tuple contructor that can take in a R named list and return Tuple object
}
\examples{
\dontrun{
t <- tuple("title"="This is title",  "brand" = "Product brand")
}
}

