% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/export-data.R
\name{write_csv}
\alias{write_csv}
\title{Export qtable to csv}
\usage{
write_csv(table_obj, ...)
}
\arguments{
\item{table_obj}{qtable object to be written to csv file.}

\item{...}{Relevant parameters to \code{\link{write.csv}}}
}
\description{
write_csv exports qtable object to csv file. It uses \code{\link{write.csv}}
to write to csv file.
}
\examples{
\\donotrun{
write_csv(qtable_obj, file = "qtable.csv")
}
}

