% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/QClasses.R
\name{QTable}
\alias{QTable}
\title{Constructor function for QTable
Extends data.frame to contain meta data such as Id attribute}
\usage{
QTable(..., genId = FALSE)
}
\description{
Constructor function for QTable
Extends data.frame to contain meta data such as Id attribute
}

